+++ pkg/analyzer/test/src/diagnostics/test_all.dart
@@ -443,6 +443,8 @@ import 'missing_enum_constant_in_switch_test.dart'
     as missing_enum_constant_in_switch;
 import 'missing_exception_value_test.dart' as missing_exception_value;
 import 'missing_field_type_in_struct_test.dart' as missing_field_type_in_struct;
+import 'missing_override_of_must_be_overridden_test.dart'
+    as missing_override_of_must_be_overridden;
 import 'missing_required_param_test.dart' as missing_required_param;
 import 'missing_return_test.dart' as missing_return;
 import 'missing_size_annotation_carray_test.dart'
@@ -1087,6 +1089,7 @@ main() {
     missing_enum_constant_in_switch.main();
     missing_exception_value.main();
     missing_field_type_in_struct.main();
+    missing_override_of_must_be_overridden.main();
     missing_required_param.main();
     missing_return.main();
     missing_size_annotation_carray.main();