+++ pkg/dart2wasm/lib/compile.dart
@@ -45,7 +45,7 @@ Future<Uint8List?> compileToModule(compiler.CompilerOptions options,
   CompilerOptions compilerOptions = CompilerOptions()
     ..target = target
     ..sdkRoot = options.sdkPath
-    ..environmentDefines = {}
+    ..environmentDefines = options.environment
     ..verbose = false
     ..onDiagnostic = diagnosticMessageHandler
     ..nnbdMode = NnbdMode.Strong;