+++ runtime/tests/vm/dart/use_flag_test_helper.dart
@@ -24,6 +24,7 @@ final genSnapshot = File(path.join(buildDir, _genSnapshotBase)).existsSync()
     : path.join(buildDir + '_X64', _genSnapshotBase);
 final aotRuntime = path.join(
     buildDir, 'dart_precompiled_runtime' + (Platform.isWindows ? '.exe' : ''));
+final isSimulator = path.basename(buildDir).contains('SIM');
 
 String? get clangBuildToolsDir {
   String archDir;
@@ -49,7 +50,10 @@ Future<void> assembleSnapshot(String assemblyPath, String snapshotPath,
   String cc = 'gcc';
   String shared = '-shared';
 
-  if (buildDir.endsWith('SIMARM') || buildDir.endsWith('SIMARM64')) {
+  if (buildDir.endsWith('SIMRISCV64')) {
+    cc = 'riscv64-linux-gnu-gcc';
+  } else if (isSimulator) {
+    // For other simulators, depend on buildtools existing.
     final clangBuildTools = clangBuildToolsDir;
     if (clangBuildTools != null) {
       cc = path.join(clangBuildTools, 'clang');
@@ -61,6 +65,7 @@ Future<void> assembleSnapshot(String assemblyPath, String snapshotPath,
     cc = 'clang';
   }
 
+  // TODO(49519): What do we need to change for SIMRISCV64?
   if (buildDir.endsWith('SIMARM')) {
     ccFlags.add('--target=armv7-linux-gnueabihf');
   } else if (buildDir.endsWith('SIMARM64')) {
@@ -102,9 +107,7 @@ Future<void> stripSnapshot(String snapshotPath, String strippedPath,
 
   var strip = 'strip';
 
-  if (buildDir.endsWith('SIMARM') ||
-      buildDir.endsWith('SIMARM64') ||
-      (Platform.isMacOS && forceElf)) {
+  if (isSimulator || (Platform.isMacOS && forceElf)) {
     final clangBuildTools = clangBuildToolsDir;
     if (clangBuildTools != null) {
       strip = path.join(clangBuildTools, 'llvm-strip');